Refactor seqwall/
