+++ pkg/seqwall/placeholders.go
@@ -1,0 +1,3 @@
+package seqwall
+
+const CurrentMigrationPlaceholder = "{current_migration}"
